Add subdirectories for mobile file work
Server is our server, client is specifically the web client that a user sees when they visit our server. To avoid deploying our mobile code to Heroku we create ios and android folders (which will then contain all the various config files for these projects), and put these folders in the .slugignore.