Refactor lote create controller to return lote
A bit hackish with regards to timestamps, but this refactors lote create to return an object of the same structure as a lote we recieve when querying the lotes table. This can then be passed into things such as sockets to avoid another database query for the values we literally just inserted.