render icons for lotes on map pages
Since lotes are tied to location we want to be able to view them on a map. We have the option to open one in mapview, this is the beginning of work to have full featured interaction with lotes in the map. Users should be able to select one lote and see it's message, potentially see radius.

We distinguish users lotes from each other by hasing their email address to a color from the material ui color palette, I am very proud of this even though it's a library because I only found the library after deciding that this is what we should do, and HEY, it already exists!